+++ PGLib/include/util/MeshTracker.hpp
@@ -10,6 +10,7 @@
 #include <filesystem>
 #include <functional>
 #include <memory>
+#include <mutex>
 #include <string>
 #include <unordered_map>
 #include <unordered_set>
@@ -64,6 +65,21 @@ private:
 
     using AltTex3DIndices = std::unordered_set<unsigned int>;
 
+    struct PathSizeHash {
+        auto operator()(const std::pair<std::filesystem::path, size_t>& key) const noexcept -> size_t
+        {
+            const size_t h1 = std::hash<std::wstring> {}(key.first.wstring());
+            const size_t h2 = std::hash<size_t> {}(key.second);
+
+            // standard hash combine
+            return h1
+                ^ (h2 + 0x9e3779b97f4a7c15ULL + (h1 << 6) + (h1 >> 2)); // NOLINT(cppcoreguidelines-avoid-magic-numbers)
+        }
+    };
+    static inline std::mutex s_otherWeightVariantsMutex;
+    static inline std::unordered_map<std::pair<std::filesystem::path, size_t>, nifly::NifFile, PathSizeHash>
+        s_otherWeightVariants;
+
 public:
     MeshTracker(const std::filesystem::path& origMeshPath);
 
@@ -71,8 +87,8 @@ public:
     void load();
     void load(const std::shared_ptr<nifly::NifFile>& origNifFile, const unsigned long long& origCrc32);
     auto stageMesh() -> nifly::NifFile*;
-    auto commitBaseMesh() -> bool;
-    auto commitDupMesh(const FormKey& formKey,
+    auto commitBaseMesh(bool isWeighted) -> bool;
+    auto commitDupMesh(const FormKey& formKey, bool isWeighted,
         const std::unordered_map<unsigned int, NIFUtil::TextureSet>& altTexResults,
         const std::unordered_set<unsigned int>& nonAltTexShapes) -> bool;
 
@@ -81,10 +97,15 @@ public:
 
     auto saveMeshes() -> std::pair<std::vector<MeshResult>, std::pair<unsigned long long, unsigned long long>>;
 
+    static void validateWeightedVariants();
+
 private:
+    void processWeightVariant();
+
     // Helpers
     static auto compareMesh(const nifly::NifFile& meshA, const nifly::NifFile& meshB,
-        const std::unordered_set<unsigned int>& enforceCheckShapeTXSTA, bool compareAllTXST = false) -> bool;
+        const std::unordered_set<unsigned int>& enforceCheckShapeTXSTA, bool compareAllTXST = false,
+        bool skipVertCheck = false) -> bool;
 
     static auto compareBSTriShape(const nifly::BSTriShape& shapeA, const nifly::BSTriShape& shapeB) -> bool;
 
@@ -107,4 +128,6 @@ private:
     static auto getComparableBlocks(const nifly::NifFile* nif) -> std::vector<nifly::NiObject*>;
 
     static auto get3dIndices(const nifly::NifFile* nif) -> std::unordered_map<nifly::NiObject*, int>;
+
+    static auto getOtherWeightVariant(const std::filesystem::path& nifPath) -> std::filesystem::path;
 };