+++ PGLib/src/util/MeshTracker.cpp
@@ -20,6 +20,7 @@
 #include <fstream>
 #include <ios>
 #include <memory>
+#include <mutex>
 #include <sstream>
 #include <stdexcept>
 #include <string>
@@ -78,7 +79,7 @@ auto MeshTracker::stageMesh() -> nifly::NifFile*
     return m_stagedMeshPtr;
 }
 
-auto MeshTracker::commitBaseMesh() -> bool
+auto MeshTracker::commitBaseMesh(bool isWeighted) -> bool
 {
     if (!m_outputMeshes.empty()) {
         // Base mesh already committed
@@ -101,6 +102,11 @@ auto MeshTracker::commitBaseMesh() -> bool
         return false;
     }
 
+    if (isWeighted) {
+        // Process weighted variant
+        processWeightVariant();
+    }
+
     const MeshResult meshResult = { .meshPath = {}, .altTexResults = {}, .idxCorrections = {} };
 
     nifly::NifFile newMesh;
@@ -120,7 +126,7 @@ auto MeshTracker::commitBaseMesh() -> bool
     return true;
 }
 
-auto MeshTracker::commitDupMesh(const FormKey& formKey,
+auto MeshTracker::commitDupMesh(const FormKey& formKey, bool isWeighted,
     const std::unordered_map<unsigned int, NIFUtil::TextureSet>& altTexResults,
     const std::unordered_set<unsigned int>& nonAltTexShapes) -> bool
 {
@@ -165,6 +171,11 @@ auto MeshTracker::commitDupMesh(const FormKey& formKey,
         m_stagedMesh.Clear();
 
         return false;
+    }
+
+    if (isWeighted) {
+        // Process weighted variant
+        processWeightVariant();
     }
 
     // Add new mesh
@@ -294,12 +305,47 @@ auto MeshTracker::saveMeshes() -> pair<vector<MeshResult>, pair<unsigned long lo
     return { output, { m_origCrc32, 0 } };
 }
 
+void MeshTracker::validateWeightedVariants()
+{
+    const std::scoped_lock lock(s_otherWeightVariantsMutex);
+    for (const auto& [key, nifFile] : s_otherWeightVariants) {
+        Logger::error(L"Weighted mesh variant for '{}' was not created. This is an issue with the original plugins and "
+                      L"can cause CTDs.",
+            key.first.wstring());
+    }
+    s_otherWeightVariants.clear();
+}
+
+void MeshTracker::processWeightVariant()
+{
+    // Check other weight variant cache
+    const std::scoped_lock lock(s_otherWeightVariantsMutex);
+    const auto dupIdx = m_outputMeshes.size();
+    // check if other variant exists
+    const auto otherVariantPath = getOtherWeightVariant(m_origMeshPath);
+    if (s_otherWeightVariants.contains({ otherVariantPath, dupIdx })) {
+        if (!compareMesh(m_stagedMesh, s_otherWeightVariants[{ otherVariantPath, dupIdx }], {}, true, true)) {
+            // different from each other, post error
+            Logger::error(L"Weighted mesh variant for '{}' differs from other weight variant '{}'. This is an "
+                          L"issue with the original models or bad pbr json definitions and can cause CTDs.",
+                m_origMeshPath.wstring(), otherVariantPath.wstring());
+        }
+
+        // delete from cache to free memory
+        s_otherWeightVariants.erase({ otherVariantPath, dupIdx });
+    } else {
+        // add to cache
+        s_otherWeightVariants[{ m_origMeshPath, dupIdx }] = nifly::NifFile();
+        s_otherWeightVariants[{ m_origMeshPath, dupIdx }].CopyFrom(m_stagedMesh);
+    }
+}
+
 //
 // ANY changes in patchers that involve WRITING new properties must be included in the equality operators below
 //
 
 auto MeshTracker::compareMesh(const nifly::NifFile& meshA, const nifly::NifFile& meshB,
-    const std::unordered_set<unsigned int>& enforceCheckShapeTXSTA, bool compareAllTXST) -> bool
+    const std::unordered_set<unsigned int>& enforceCheckShapeTXSTA, bool compareAllTXST, bool skipVertCheck) -> bool
 {
     // This should be compared before sorting blocks (sorting blocks should happen last)
     const auto blocksA = getComparableBlocks(&meshA);
@@ -391,7 +437,7 @@ auto MeshTracker::compareMesh(const nifly::NifFile& meshA, const nifly::NifFile&
             }
             if (bstrishapeA != nullptr && bstrishapeB != nullptr) {
                 // compare trishape helper
-                if (!compareBSTriShape(*bstrishapeA, *bstrishapeB)) {
+                if (!skipVertCheck && !compareBSTriShape(*bstrishapeA, *bstrishapeB)) {
                     return false;
                 }
             }
@@ -485,6 +531,16 @@ auto MeshTracker::compareMesh(const nifly::NifFile& meshA, const nifly::NifFile&
 
 auto MeshTracker::compareBSTriShape(const nifly::BSTriShape& shapeA, const nifly::BSTriShape& shapeB) -> bool
 {
+    if (!shapeA.HasVertexColors() && !shapeB.HasVertexColors()) {
+        // nothing to check
+        return true;
+    }
+
+    if (shapeA.HasVertexColors() != shapeB.HasVertexColors()) {
+        // only one shape has vertex colors
+        return false;
+    }
+
     const auto vertdataA = shapeA.vertData;
     const auto vertdataB = shapeB.vertData;
     if (vertdataA.size() != vertdataB.size()) {
@@ -699,4 +755,22 @@ auto MeshTracker::get3dIndices(const nifly::NifFile* nif) -> unordered_map<nifly
     }
 
     return blocks;
+}
+
+auto MeshTracker::getOtherWeightVariant(const std::filesystem::path& nifPath) -> std::filesystem::path
+{
+    // convert m_origMeshPath to weight slider variant
+    std::filesystem::path weightVariant = nifPath;
+
+    const static wstring oneWeightVariant = L"_1.nif";
+    const static wstring zeroWeightVariant = L"_0.nif";
+
+    const std::wstring nifPathStr = nifPath.wstring();
+    if (nifPathStr.ends_with(oneWeightVariant)) {
+        weightVariant = nifPathStr.substr(0, nifPathStr.size() - oneWeightVariant.size()) + zeroWeightVariant;
+    } else if (nifPathStr.ends_with(zeroWeightVariant)) {
+        weightVariant = nifPathStr.substr(0, nifPathStr.size() - zeroWeightVariant.size()) + oneWeightVariant;
+    }
+
+    return weightVariant;
 }